CompactLines: Refrain from generating JSX links
The transformation steps embed messages in ParseResult mostly for use
in DebugView, where they are displayed to the user. CompactLines puts
in JSX elements, which require React. This has to be removed so that
`lib/` can be made into pure JS.

Knock out the aforementioned logic, limiting CompactLines' messages to
reporting how many footnotes are generated, not links that jump to the
page containing the footnotes